Move KeyMgmt class into its own file; clean up some db_lib code
